converted std.py to std/ to break up code into separate files
