+++ tools/python/jsont/transforms/std/ss.json
